cleanup; restructuring state items from icq to state.rs
