+++ contracts/orbital-core/src/contract.rs
@@ -1,8 +1,11 @@
 use crate::{
     admin_logic::admin,
-    icq::{self, RecipientTxsResponse, RECIPIENT_TXS, TRANSFERS},
-    msg::GetTransfersAmountResponse,
-    state::{ClearingAccountConfig, OrbitalDomainConfig, UserConfig, USER_NONCE},
+    icq::{self},
+    msg::{GetTransfersAmountResponse, RecipientTxsResponse},
+    state::{
+        ClearingAccountConfig, OrbitalDomainConfig, UserConfig, RECIPIENT_TXS, TRANSFERS,
+        USER_NONCE,
+    },
     user_logic::user,
     utils::{extract_ica_identifier_from_port, get_ica_identifier, OpenAckVersion},
 };
