+++ contracts/orbital-core/src/state.rs
@@ -1,6 +1,8 @@
 use cosmwasm_schema::cw_serde;
 use cosmwasm_std::{Addr, Uint64};
 use cw_storage_plus::{Item, Map};
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
 
 /// keeping track of registered user IDs which get incremented
 /// with each new registration. it's needed to generate unique
@@ -20,6 +22,19 @@ pub const ORBITAL_DOMAINS: Map<String, OrbitalDomainConfig> = Map::new("domains"
 /// - Some: clearing account has been registered and is ready for use
 pub const CLEARING_ACCOUNTS: Map<String, Option<ClearingAccountConfig>> =
     Map::new("clearing_accounts");
+
+/// contains all transfers mapped by a recipient address observed by the contract.
+pub const RECIPIENT_TXS: Map<String, Vec<Transfer>> = Map::new("recipient_txs");
+/// contains number of transfers to addresses observed by the contract.
+pub const TRANSFERS: Item<u64> = Item::new("transfers");
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
+pub struct Transfer {
+    pub recipient: String,
+    pub sender: String,
+    pub denom: String,
+    pub amount: String,
+}
 
 #[cw_serde]
 pub struct ClearingAccountConfig {
